// web/platform-admin-web/src/api/generated/model/createResourceRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { CreateResourceRequestDtoActionsItem } from './createResourceRequestDtoActionsItem';
import type { CreateResourceRequestDtoScope } from './createResourceRequestDtoScope';
import type { CreateResourceRequestDtoType } from './createResourceRequestDtoType';

export interface CreateResourceRequestDto {
  /** 리소스 코드 */
  code: string;
  /** 리소스 이름 */
  name: string;
  /** 리소스 유형 */
  type: CreateResourceRequestDtoType;
  /** 리소스 관리 범위 */
  scope: CreateResourceRequestDtoScope;
  /**
     * 부모 리소스 식별자
     * @nullable
     */
  parent: string | null;
  /**
     * 리소스 경로
     * @nullable
     */
  path: string | null;
  /**
     * 아이콘
     * @nullable
     */
  icon: string | null;
  /**
     * 정렬 순서
     * @nullable
     */
  sortOrder: number | null;
  /** 리소스 액션 목록 */
  actions: CreateResourceRequestDtoActionsItem[];
}