// web/platform-admin-web/src/api/generated/model/getResourceListFiltersDtoScope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 리소스 관리 범위
 */
export type GetResourceListFiltersDtoScope = typeof GetResourceListFiltersDtoScope[keyof typeof GetResourceListFiltersDtoScope];


export const GetResourceListFiltersDtoScope = {
  PLATFORM: 'PLATFORM',
  ORGANIZATION: 'ORGANIZATION',
} as const;