// web/platform-admin-web/src/api/generated/model/getResourceListItemType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 리소스 유형
 */
export type GetResourceListItemType = typeof GetResourceListItemType[keyof typeof GetResourceListItemType];


export const GetResourceListItemType = {
  MENU: 'MENU',
  COMPONENT: 'COMPONENT',
} as const;