// web/platform-admin-web/src/api/generated/model/updateResourceSortRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { UpdateResourceSortItemDto } from './updateResourceSortItemDto';

export interface UpdateResourceSortRequestDto {
  /** 정렬 대상 목록 */
  items: UpdateResourceSortItemDto[];
}